refactor: EmptySections와 FilledSections 분리
